refactor: 重组Rust包结构
