+++ tauri-app/AGENTS.md
@@ -31,32 +31,57 @@ src/
 - i18n via `vue-i18n`; add strings to both `src/shared/locales/en.json` and `zh.json`
 - Path alias: `@/` maps to `src/` (configured in `vite.config.ts` and `tsconfig.json`)
 
-### Backend (Rust, Cargo workspace)
+### Backend (Rust, single crate)
 
 ```
 src-tauri/              # Main Tauri app crate (micyou-app)
   src/
-    commands/           # Tauri invoke commands
-    lib.rs              # Tauri plugin registration
+    protocol/           # Protobuf definitions + wire constants (PACKET_MAGIC, PORT, ...)
+    audio/              # Audio DSP + output pipeline (was micyou-audio crate)
+      settings.rs       # AudioDspSettings, EqualizerConfig
+      engine.rs         # AudioOutputManager, RubatoResampler
+      dsp.rs            # DspProcessor: chain orchestrator + spectrum
+      equalizer.rs      # 10-band biquad peaking EQ
+      resampler.rs      # Adaptive playback-rate resampler
+      agc.rs            # Automatic gain control
+      vad.rs            # Voice activity detection (RMS threshold)
+      dereverb.rs       # Delay-line comb filter dereverb
+      aec.rs            # AEC7 acoustic echo cancellation (ONNX)
+      jitter_buffer.rs  # Packet reordering + FEC recovery
+      pipeline.rs       # Audio output thread (decode → jitter → resample → DSP → device)
+      noise/            # Noise reduction implementations
+        rnnoise.rs      # RNNoise (nnnoiseless)
+        ulunas.rs       # Ulunas ONNX AI denoiser
+        speex.rs        # Speexdsp spectral subtraction
+        lightweight.rs  # Simple noise gate / expander
+    network/            # All network I/O
+      mdns.rs           # mDNS service discovery
+      discovery.rs      # LAN IP enumeration + scoring
+      tcp_server.rs     # TCP audio server (handshake + control loop)
+      udp_server.rs     # UDP audio server
+      web_server.rs     # HTTPS WebSocket server (feature-gated)
+    platform/           # OS-specific functionality
+      adb_manager.rs    # ADB device management
+      vbcable.rs        # VB-Cable installer (feature-gated)
+      blackhole.rs      # BlackHole audio device management (macOS)
+    commands/           # All #[tauri::command] handlers
+      mod.rs            # Core commands (start/stop server, mute, tray, window, ...)
+      about.rs          # About/sponsors/export-log
+    lib.rs              # Module declarations, ServerState, run()
     main.rs             # Entry point
-    tcp_server.rs       # TCP audio server
-    udp_server.rs       # UDP audio server
-    web_server.rs       # HTTPS web server (feature-gated)
-    adb_manager.rs      # ADB device management (feature-gated)
-    vbcable.rs          # VB-Cable installer (feature-gated)
-    ...
-  proto/                # (empty; protocol proto is in micyou-protocol)
-crates/
-  micyou-protocol/      # Protobuf protocol definitions + codegen
-    proto/network.proto # Protocol definition
-    build.rs            # prost codegen
-  micyou-audio/         # Audio DSP crate
+    stats.rs            # NetworkStats (atomic counters)
+    tray.rs             # System tray
+  proto/
+    network.proto       # Protobuf schema
+  build.rs              # protobuf + tauri codegen
 ```
 
 **Cargo features** (in `src-tauri/Cargo.toml`):
 - `vbcable` (default) — Windows VB-Cable installer; pulls in `zip`
 - `adb` (default) — ADB device management
 - `web-server` (default) — HTTPS web server for browser-based mic; pulls in `axum`, `rustls`
+- `dsp` (default) — Audio DSP processing; pulls in `nnnoiseless`, `ort`, `rustfft`
+- `noise-suppression` — ONNX-based AI denoising (Ulunas); requires `dsp`
 
 ## Commands
 
@@ -69,10 +94,9 @@ npm run tauri build          # Production build
 npm run sync-version         # Sync version from ../gradle.properties → tauri.conf.json, Cargo.toml, package.json
 
 # Rust workspace
-cargo build                  # Build all crates
+cargo check                  # Type-check only (fast)
+cargo build                  # Build
 cargo test                   # Run tests
-cargo build -p micyou-protocol   # Build protocol crate only
-cargo build -p micyou-audio      # Build audio crate only
 ```
 
 ## Version management
@@ -96,6 +120,6 @@ Do not edit version numbers in Tauri/Cargo/package.json directly — always chan
 - Frontend strings use `vue-i18n` (`t('key')`) with JSON locale files, not hardcoded
 - Keep features self-contained: components + composables inside each feature directory
 - Shared/reusable code goes in `src/shared/`
-- Rust code uses `prost` for protobuf codegen; `.proto` files live in `crates/micyou-protocol/proto/`
-- Tauri commands are registered in `src-tauri/src/lib.rs`
+- Rust code uses `prost` for protobuf codegen; `.proto` file in `src-tauri/proto/`
+- Tauri commands are registered in `src-tauri/src/lib.rs`, implementations in `src-tauri/src/commands/`
 - Popup windows use hash-based routing in `main.ts` (`#/popup/ip`, `#/popup/*`)
