--- tauri-app/crates/micyou-audio/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "micyou-audio"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-
-[features]
-default = ["dsp"]
-dsp = ["dep:nnnoiseless", "dep:ort", "dep:ort-tract", "dep:rustfft", "dep:ndarray", "dep:num-complex"]
-noise-suppression = ["dsp", "dep:ort", "dep:ort-tract"]
-
-[dependencies]
-micyou-protocol = { path = "../micyou-protocol" }
-serde = { version = "1", features = ["derive"] }
-log = "0.4.32"
-cpal = "0.15.3"
-rubato = "3.0.0"
-ringbuf = "0.3.3"
-
-# DSP dependencies (optional)
-nnnoiseless = { version = "0.5.2", optional = true }
-ort = { version = "=2.0.0-rc.12", default-features = false, features = ["std", "ndarray", "api-17", "alternative-backend"], optional = true }
-ort-tract = { version = "0.3.0", optional = true }
-rustfft = { version = "6.4.1", optional = true }
-ndarray = { version = "0.16", optional = true }
-num-complex = { version = "0.4", optional = true }
