--- tauri-app/crates/micyou-protocol/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "micyou-protocol"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-
-[dependencies]
-prost = "0.14.3"
-bytes = "1.11.1"
-serde = { version = "1", features = ["derive"] }
-
-[build-dependencies]
-prost-build = "0.14.3"
-protoc-bin-vendored = "3.2.0"
