--- tauri-app/crates/micyou-protocol/build.rs
@@ -1,5 +1,0 @@
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    std::env::set_var("PROTOC", protoc_bin_vendored::protoc_bin_path().unwrap());
-    prost_build::compile_protos(&["proto/network.proto"], &["proto/"])?;
-    Ok(())
-}
