+++ tauri-app/src-tauri/Cargo.toml
@@ -10,16 +10,14 @@ name = "tauri_app_lib"
 crate-type = ["staticlib", "cdylib", "rlib"]
 
 [features]
-default = ["vbcable", "adb", "web-server"]
+default = ["vbcable", "adb", "web-server", "dsp"]
 vbcable = ["dep:zip"]
 adb = []
 web-server = ["dep:axum", "dep:rcgen", "dep:rustls", "dep:tokio-rustls", "dep:rustls-pemfile"]
+dsp = ["dep:nnnoiseless", "dep:ort", "dep:ort-tract", "dep:rustfft", "dep:ndarray", "dep:num-complex"]
+noise-suppression = ["dsp", "dep:ort", "dep:ort-tract"]
 
 [dependencies]
-# Workspace crates
-micyou-protocol = { path = "../crates/micyou-protocol" }
-micyou-audio = { path = "../crates/micyou-audio", features = ["dsp"] }
-
 # Tauri
 tauri = { version = "2", features = ["macos-private-api", "tray-icon", "image-png"] }
 tauri-plugin-opener = "2"
@@ -35,10 +33,16 @@ tokio = { version = "1.52.3", features = ["full"] }
 tokio-util = "0.7.18"
 log = "0.4.32"
 bytes = "1.11.1"
+
+# Protobuf
 prost = "0.14.3"
 
 # Audio device enumeration
 cpal = "0.15.3"
+
+# Audio processing
+rubato = "3.0.0"
+ringbuf = "0.3.3"
 
 # Networking
 mdns-sd = "0.20.0"
@@ -60,8 +64,18 @@ rustls = { version = "0.23", optional = true }
 tokio-rustls = { version = "0.26", optional = true }
 rustls-pemfile = { version = "2", optional = true }
 
+# DSP dependencies (optional)
+nnnoiseless = { version = "0.5.2", optional = true }
+ort = { version = "=2.0.0-rc.12", default-features = false, features = ["std", "ndarray", "api-17", "alternative-backend"], optional = true }
+ort-tract = { version = "0.3.0", optional = true }
+rustfft = { version = "6.4.1", optional = true }
+ndarray = { version = "0.16", optional = true }
+num-complex = { version = "0.4", optional = true }
+
 [build-dependencies]
 tauri-build = { version = "2", features = [] }
+prost-build = "0.14.3"
+protoc-bin-vendored = "3.2.0"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3", features = ["winsock2"] }
