+++ tauri-app/src-tauri/src/audio/process/mod.rs
@@ -1,0 +1,9 @@
+pub mod settings;
+pub mod noise;
+pub mod equalizer;
+pub mod resampler;
+pub mod dereverb;
+pub mod agc;
+pub mod vad;
+pub mod dsp;
+pub mod aec;
