+++ tauri-app/src-tauri/src/audio/process/noise/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "dsp")]
+pub mod rnnoise;
+#[cfg(feature = "noise-suppression")]
+pub mod ulunas;
+#[cfg(feature = "dsp")]
+pub mod speex;
+pub mod lightweight;
