+++ tauri-app/src-tauri/src/main.rs
@@ -1,6 +1,6 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {
-    micyou_audio::init_onnx_runtime();
+    tauri_app_lib::audio::init_onnx_runtime();
     tauri_app_lib::run()
 }
