+++ tauri-app/src-tauri/src/platform/adb_manager.rs
