+++ tauri-app/src-tauri/src/platform/mod.rs
@@ -1,0 +1,3 @@
+pub mod adb_manager;
+pub mod vbcable;
+pub mod blackhole;
